+++ trip-helper/src/definition/storage/IuserPreferenceStorage.ts
@@ -1,0 +1,7 @@
+import { IPreference } from "../helper/userPreference";
+
+export interface IuserPreferenceStorage {
+    storeUserPreference(preference: IPreference): Promise<void>;
+    getUserPreference(): Promise<IPreference | null>;
+    clearUserPreference(): Promise<void>;
+}
